+++ example/package.json
@@ -16,6 +16,7 @@
     "@types/node-static": "^0.7.6",
     "@vitejs/plugin-vue": "^1.1.5",
     "@vue/compiler-sfc": "^3.0.7",
+    "@vueuse/core": "^4.3.4",
     "cross-env": "^7.0.3",
     "node-static": "^0.7.11",
     "typescript": "^4.2.3",
